Added AMR decoding - all recorded audio can now be added to movies
As there are no open source java AMR decoding libraries, this
capability is proviced via a native armeabi extension (opencore-amr).
The compiled library (opencore-amrnb-wrapper.so) is in the repository,
but the wrapper source and a build file are also included, as these do
not seem to be available anywhere else. Compiling the native library
requires the Android NDK and the opencore-amr source (see build.sh).